+++ apps/nextjs/src/app/[locale]/_client-providers/trpc.tsx
@@ -18,7 +18,8 @@ import superjson from "superjson";
 import type { SuperJSONResult } from "superjson";
 
 import type { AppRouter } from "@homarr/api";
-import { clientApi, createHeadersCallbackForSource, getTrpcUrl } from "@homarr/api/client";
+import { clientApi, getTrpcUrl } from "@homarr/api/client";
+import { createHeadersCallbackForSource } from "@homarr/api/shared";
 
 import { env } from "~/env.mjs";
