+++ packages/api/package.json
@@ -8,7 +8,8 @@
     ".": "./src/index.ts",
     "./client": "./src/client.ts",
     "./server": "./src/server.ts",
-    "./websocket": "./src/websocket.ts"
+    "./websocket": "./src/websocket.ts",
+    "./shared": "./src/shared.ts"
   },
   "main": "./index.ts",
   "types": "./index.ts",
